Big update before making the sound board more publicly visible through the Dialed video comment section.
Added sounds for Ep29, 30, 31 and 36.

Added a head shot of our long suffering hero Jordi Cortes.

Added more details about the sound board at the bottom of the page.

Moved the head shot images to their own folder.

Updated the CSS for a better layout. The head shots still scale a bit weirdly but it'll do for now.